# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SMRITI-OS - Sovereign Retail OS",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/barcodes/scan/{barcode}": {
            "get": {
                "tags": [
                    "barcode"
                ],
                "summary": "Scan Barcode",
                "description": "HOT PATH: Resolve barcode / stockno to item details + live stock.\nIn Shoper9, the stockno IS the barcode. This endpoint resolves both.\nTarget: < 50ms via indexed stockno lookup.",
                "operationId": "scan_barcode_api_v1_barcodes_scan__barcode__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "barcode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Barcode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/barcodes/scan-multi": {
            "get": {
                "tags": [
                    "barcode"
                ],
                "summary": "Scan Multiple",
                "description": "Resolve multiple barcodes/stocknos in a single call.\nUseful for bulk validation before billing.",
                "operationId": "scan_multiple_api_v1_barcodes_scan_multi_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "barcodes",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Barcodes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/barcodes/item-info/{stockno}": {
            "get": {
                "tags": [
                    "barcode"
                ],
                "summary": "Get Item Info",
                "description": "Full item info from Itemmaster + Stockmaster for a given stockno.",
                "operationId": "get_item_info_api_v1_barcodes_item_info__stockno__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "stockno",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Stockno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/barcodes/print": {
            "post": {
                "tags": [
                    "barcode"
                ],
                "summary": "Print Barcode",
                "description": "Print barcode label for a stockno.\nResolves item details from Itemmaster for label data.",
                "operationId": "print_barcode_api_v1_barcodes_print_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "additionalProperties": true,
                                "type": "object",
                                "title": "Payload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}